[MOV] stock_account: Re-organized directory structure.
