+++ addons/stock_account/data/product_data.xml
