+++ addons/stock_account/data/stock_account_data.xml
