+++ addons/stock_account/data/stock_account_demo.yml
