+++ addons/stock_account/data/stock_account_post_install.yml
