+++ addons/stock_account/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-import models
-import wizard
+import product
+import res_config
+import stock_account
