+++ addons/stock_account/models/product.py
