+++ addons/stock_account/models/res_config.py
