+++ addons/stock_account/models/stock_account.py
