+++ addons/stock_account/views/product_views.xml
