+++ addons/stock_account/views/res_config_views.xml
