+++ addons/stock_account/views/stock_account_views.xml
